requirement.txt
alembic==1.7.6
allure-pytest==2.8.40
allure-python-commons==2.8.40
apturl==0.5.2
astor==0.8.1
attrs==21.4.0
Authlib==0.15.5
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
bcrypt==3.1.7
beautifulsoup4==4.10.0
bidict==0.21.4
blinker==1.4
boto==2.49.0
Brlapi==0.7.0
Brotli==1.0.9
certifi==2019.11.28
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.4
click==8.0.4
colorama==0.4.3
command-not-found==0.3
commonmark==0.9.1
cryptography==3.4.8
cssselect==1.1.0
cupshelpers==1.0
dbus-python==1.2.16
decorator==5.0.9
defer==1.0.6
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.2.0
duplicity==0.8.12.0
email-validator==1.1.3
entrypoints==0.3
et-xmlfile==1.1.0
eventlet==0.33.0
execnet==1.9.0
fasteners==0.14.1
filelock==3.4.2
Flask==2.0.3
Flask-Babel==2.0.0
Flask-Compress==1.11
Flask-Gravatar==0.5.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.3
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
future==0.18.2
greenlet==1.1.2
httpagentparser==1.9.2
httplib2==0.14.0
idna==2.8
importlib-metadata==4.11.2
importlib-resources==5.4.0
iniconfig==1.1.1
ipdb==0.13.9
ipython==7.27.0
itsdangerous==2.1.0
jedi==0.18.0
Jinja2==3.0.3
jproperties==2.1.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
ldap3==2.9.1
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
monotonic==1.5
more-itertools==8.8.0
MouseInfo==0.1.3
msedge-selenium-tools==3.141.3
netifaces==0.10.4
nose==1.3.7
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.9
packaging==21.3
pandas==1.3.5
parameterized==0.8.1
paramiko==2.9.2
parso==0.8.3
passlib==1.7.4
pdfminer.six==20201018
pexpect==4.6.0
pgadmin4==6.5
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==2.4.1
pluggy==1.0.0
prompt-toolkit==3.0.28
protobuf==3.6.1
psutil==5.9.0
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
PyAutoGUI==0.9.53
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
PyGetWindow==0.0.9
Pygments==2.11.2
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyMySQL==1.0.2
PyNaCl==1.3.0
pyOpenSSL==20.0.1
pyotp==2.6.0
pyparsing==3.0.6
pyperclip==1.8.2
PyRect==0.1.4
pyRFC3339==1.1
PyScreeze==0.1.28
pytest==6.2.5
pytest-ast-transformer==1.0.3
pytest-forked==1.4.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-ordering==0.6
pytest-rerunfailures==10.1
pytest-soft-assertions==0.1.2
pytest-xdist==2.5.0
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-engineio==4.3.1
python-socketio==5.5.2
python3-xlib==0.15
pytweening==1.0.4
pytz==2021.3
pyxdg==0.26
PyYAML==5.3.1
qrcode==7.3.1
reportlab==3.5.34
requests==2.25.1
requests-unixsocket==0.2.0
rich==10.9.0
sbvirtualdisplay==1.0.0
SecretStorage==2.3.1
selenium==3.141.0
setuptools-scm==6.0.1
simplejson==3.16.0
six==1.14.0
sortedcontainers==2.4.0
soupsieve==2.3.1
speaklater3==1.4
SQLAlchemy==1.4.31
sqlparse==0.4.2
sshtunnel==0.4.0
systemd-python==234
texttable==1.6.4
toml==0.10.2
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
ua-parser==0.10.0
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
user-agents==2.2.0
virtualenv==20.13.1
wadllib==1.3.3
wcwidth==0.2.5
webdriver-manager==3.5.4
Werkzeug==2.0.3
WTForms==3.0.1
xkit==0.0.0
zipp==3.7.0
